requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-openai
openai
langchain-chroma
chromadb
pymupdf
streamlit
pandas
pyyaml